% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.r
\name{balanced.folds}
\alias{balanced.folds}
\title{Get balanced folds where each fold has close to overall class ratio. This only works for non-numeric inputs.}
\usage{
balanced.folds(y, nfolds = 10)
}
\arguments{
\item{y}{Vector of classes; should be a factor, logical, or character}

\item{nfolds}{Number of folds. -1 means leave-one-out}
}
\value{
Numeric vector of fold indices, one for each element of \code{y}
}
\description{
Get balanced folds where each fold has close to overall class ratio. This only works for non-numeric inputs.
}

