% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.r
\name{lars.cross.validation}
\alias{lars.cross.validation}
\title{Runs cross-validation of LARS regression}
\usage{
lars.cross.validation(x, y, nfolds = 10, folds = NULL, verbose = FALSE,
  ...)
}
\arguments{
\item{x}{A matrix of numeric predictors}

\item{y}{A factor or logical of responses}

\item{nfolds}{number of cross-validation folds (nfolds==-1 means leave-one-out)}

\item{folds}{optional, if not \code{NULL} then uses these fold assignments (one fold index for each row in \code{x})}

\item{verbose}{Use verbose output}

\item{...}{additional parameters for lars}
}
\value{
:
 \item{y}{Observed values}
 \item{predicted}{CV predicted values}
 \item{fractions}{Vector of LASSO fractions used across folds}
 \item{best.fraction}{Best LASSO fraction used}
 \item{MSE}{Mean squared error}
 \item{RMSE}{Root mean squared error}
 \item{nfolds}{Number of folds used (or -1 for leave-one-out)}
 \item{params}{List of additional parameters, if any}
 \item{coefficients}{Feature-by-fold matrix of coefficients of features across folds}
 \item{final.model}{Final lars model trained on whole data set}
}
\description{
Runs cross-validation of LARS regression
}
\examples{
# generate fake data: 50 samples, 20 variables
x <- matrix(rnorm(2000),100)
rownames(x) <- sprintf('Sample\%02d',1:100)
colnames(x) <- sprintf('Feature\%02d',1:20)
# generate fake response variable, function of columns 1-5 plus noise
y <- rowSums(sweep(x[,1:5],2,runif(5),'*')) + rnorm(100,sd=.5)
names(y) <- rownames(x)

# 10-fold cross validation with predictions
res.lars <- lars.cross.validation(x,y)
# plot predictions
plot(res.lars$y, res.lars$predicted, xlab="Observed", ylab="Predicted (hold-out)")

# correlation
cat('Pearson correlation of predictions with truth?\\n')
print(cor.test(res.lars$y, res.lars$predicted))

# plot fraction of times each variable was included in a fold
sorted.importances <- sort(rowMeans(res.lars$coefficients > 0), decreasing=TRUE)
# bar plot of top ten variables
par(mar=c(4.5,6,2,1),las=2) # larger margin on left, perpendicular axis labels
barplot(rev(sorted.importances[1:10]),horiz=TRUE, xlab='Frequency of selection')

}

