% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.r
\name{probabilities.to.ROC}
\alias{probabilities.to.ROC}
\title{Performs ROC analysis (and plot) on probability vector}
\usage{
probabilities.to.ROC(p, y, step.size = 0.01, plot = TRUE, ...)
}
\arguments{
\item{p}{Vector or probabilities of TRUE}

\item{y}{Logical, factor, or character of TRUE/FALSE}

\item{step.size}{Increments at which to calculate false and true positive rates}

\item{plot}{If \code{TRUE} then make a simple ROC curve plot.}

\item{...}{Additional parameters to pass to \code{plot} function}
}
\value{
\item{fprs}{Vector of false positive rates}
 \item{tprs}{Vector of true positive rates}
 \item{auc}{Area under the curve}
}
\description{
Performs ROC analysis (and plot) on probability vector
}

