% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.r
\name{rf.cross.validation}
\alias{rf.cross.validation}
\title{runs cross-validation Random Forests classification}
\usage{
rf.cross.validation(x, y, nfolds = 10, folds = NULL, verbose = FALSE, ...)
}
\arguments{
\item{x}{A matrix of numeric predictors}

\item{y}{A factor or logical of responses}

\item{nfolds}{number of cross-validation folds (nfolds==-1 means leave-one-out)}

\item{folds}{optional, if not \code{NULL} then uses these fold assignments (one fold index for each row in \code{x})}

\item{verbose}{Use verbose output}

\item{...}{additional parameters for randomForest}
}
\value{
\item{y}{Observed values}
 \item{predicted}{CV predicted values}
 \item{probabilities}{CV predicted class probabilities (or NULL if unavailable)}
 \item{confusion.matrix}{Confusion matrix (true-by-predicted)}
 \item{nfolds}{Number of folds used (or -1 for leave-one-out)}
 \item{params}{List of additional parameters, if any}
 \item{importances}{Feature-by-fold matrix of importances of features (mean decrease in accuracy) across folds}
 \item{final.model}{Final random forests model trained on whole data set}
}
\description{
runs cross-validation Random Forests classification
}
\examples{
# generate fake data: 50 samples, 20 variables
x <- matrix(rnorm(2000),100)
rownames(x) <- sprintf('Sample\%02d',1:100)
colnames(x) <- sprintf('Feature\%02d',1:20)
# generate fake response variable, function of columns 1-5 plus noise
y <- rowSums(sweep(x[,1:5],2,runif(5),'*')) + rnorm(100,sd=.5)
# y is now binary response variable
y <- factor(y > median(y))
names(y) <- rownames(x)

# 10-fold cross validation, returning predictions
res.rf <- rf.cross.validation(x,y2)

# plot importance of top ten variables
sorted.importances <- sort(rowMeans(res.rf$importances), decreasing=TRUE)
barplot(rev(sorted.importances[1:10]),horiz=TRUE, xlab='Mean Decrease in Accuracy')

# Plot classification ROC curve
roc <- probabilities.to.ROC(res.rf$probabilities[,2], y2,plot=TRUE)

# Report ROC AUC
cat('ROC AUC was:',roc$auc,'\\n')

}

